Fix Straight Line Bug
In some case, if use draw a nodule area as a straight line align to
x-axis or y-axis. The area will be 0. And this casue a bug which make
program crash.
So in this update, we fixed this. If a new nodule is created and the
area is 0, the width or length will be expend to 1 to make area > 0.

Change the 'checked status' button color to make it more noticable.

+++ libraries/display/canvas.py
@@ -826,7 +826,7 @@ class Canvas(QtWidgets.QWidget):
             if 0 <= ua <= 1 and 0 <= ub <= 1:
                 x = x1 + ua * (x2 - x1)
                 y = y1 + ua * (y2 - y1)
-                m = QtCore.QPoint((x3 + x4) / 2, (y3 + y4) / 2)
+                m = QtCore.QPoint(int((x3 + x4) / 2), int((y3 + y4) / 2))
                 d = distance(m - QtCore.QPoint(x2, y2))
                 yield d, i, (x, y)
